rbd: consider lastSyncTimeNotFound as image not syncing
This commit modifies code to consider image as not syncing
when the lastSyncTime is not found.

Currently replication's Resync call fails if the
lastSyncTime of an image is not found.
Instead, we should treat missing lastSyncTime as
the image not being in Syncing state.

Signed-off-by: Rakshith R <rar@redhat.com>
(cherry picked from commit fa9e254f41bbf0ba1cb49e5c8abf5fdde8134438)