rbd: add schedule again during group modification
add the group snapshot schedule again after disabling/enabling
mirroring as there might be a race condition when the snapshot
scheduler might be run at the same time when the group is disabled
which will remove the schedule for the group.
Therefore, we should re-add the schedule to avoid such a scenario.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>