nvmeof: add the option "networkMask" in StorageClass
add the option "networkMask" in SC. This
option is intoduced to replace the listener list entry
in the SC. with this param the admin does need to know
what are GWs IPs. just needs to provide the network mask
of the deployment. one way to get it, by run
`oc get network.config/cluster -o \
 jsonpath='{.status.clusterNetwork[*].cidr}'`
then, assign the value in networkMask field,
and that's it.

Signed-off-by: gadi-didi <gadi.didi@ibm.com>